requirements.txt
openai==1.25.1
pypdf==4.2.0
python-dotenv==1.0.1
streamlit==1.34.0
